Add an extract arg to RemoteSource and uncouple from Rpm system. (#1376)
* stash extract arg for remote source

* Add an extract arg to RemoteSource and uncouple from Rpm system.

This allows any target to depend on a versioned file that is
hosted in a configured site.

This allows for extracting the download and then adds to the
product graph as a resource. This would be useful for
bootstrapping files to the classpath, although the upcoming
use case is actually something we only want to be local, so still
figuring that part out.

The remote_files tasks and subsystems should be completely broken
out of the Rpm module, will do that as a followup. This also lacks
any testing, because it predated the test infra - I hope to add
a testproject and beginning of a test suite when I write the followup.
I know that there is some error handling that is needed to catch
incongruitites and failures in the extract pipeline.

* Update some comments

* fix vt API bug

* review feedback

* add extract as a payload field

(sapling split of 9a55072fb6a5be8faa8f997472a77d5096a62411)

+++ src/python/fsqio/pants/rpmbuild/tasks/rpmbuild_task.py
@@ -27,7 +27,6 @@ from pants.util.dirutil import safe_mkdir
 from pkg_resources import resource_string
 from six import string_types
 
-from fsqio.pants.rpmbuild.subsystems.remote_source_fetcher import RemoteSourceFetcher
 from fsqio.pants.rpmbuild.targets.remote_source import RemoteSource
 from fsqio.pants.rpmbuild.targets.rpm_spec import RpmSpecTarget
 
@@ -76,11 +75,9 @@ class RpmbuildTask(Task):
              help='After invoking `rpmbuild`, commit the container state to a new image')
 
   @classmethod
-  def subsystem_dependencies(cls):
-    return super(RpmbuildTask, cls).subsystem_dependencies() + (RemoteSourceFetcher.Factory.scoped(cls),)
-
-  def __init__(self, *args, **kwargs):
-    super(RpmbuildTask, self).__init__(*args, **kwargs)
+  def prepare(cls, options, round_manager):
+    round_manager.require('remote_files')
+    super(RpmbuildTask, cls).prepare(options, round_manager)
 
   @classmethod
   def product_types(cls):
@@ -150,14 +147,23 @@ class RpmbuildTask(Task):
 
     # Copy sources to the buildroot. (TODO - unify these stanzas, they differ only in being relative vs absolute paths)
     local_sources = []
+    remote_files = self.context.products.get('remoste_files')
     for source in self._remote_source_targets(target):
-
-      remote_source = RemoteSourceFetcher.Factory.scoped_instance(self).create(source)
-      source_path = remote_source.path
-      shutil.copy(source_path, build_dir)
-      local_sources.append({
-        'basename': os.path.basename(os.path.relpath(source_path, get_buildroot())),
-      })
+      mapping = remote_files.get(source)
+      # The remote_files product is a mapping of [vt.target] => { vt.results_dir: os.listdir(vt.results_dir)}.
+      # The contents could be files or dirs, so both are handled here.
+      if mapping:
+        for dirname, filenamez in mapping.items():
+          for filename in filenamez:
+            source_path = os.path.join(dirname, filename)
+            # Allow for extracted rmeote sources.
+            if os.path.isfile(source_path):
+              shutil.copy(source_path, build_dir)
+            else:
+              shutil.copytree(source_path, build_dir)
+            local_sources.append({
+              'basename': filename,
+            })
     for source_rel_path in target.sources_relative_to_buildroot():
       shutil.copy(os.path.join(get_buildroot(), source_rel_path), build_dir)
       local_sources.append({
